cloud_functions/smart_search/requirements.txt
functions-framework==3.*
google-generativeai>=0.8.3
google-cloud-bigquery>=3.14.1
google-cloud-aiplatform>=1.71.0
vertexai>=1.71.0
flask>=3.0.0
